move MTGenericReader
